refactor(captp): breakout most trap functionality from engine
